Rename OpenERPXmlRpcProxy to OdooXmlRpcProxy
